app/requirements.txt
paramiko
cryptography
Flask
mysql-connector-python
PyMySQL
peewee
python-dateutil
librouteros